refactor: Add notification container to app layout
This commit adds a new CSS class called "notification-container" to the app layout component. This container is positioned fixed at the top right corner of the page and is used to display notifications to the user. The container has a maximum width of 100% and a margin of 2rem. This change improves the organization and presentation of notifications in the application.