refactor: Update composer.json to use PHP 8.2
This commit updates the composer.json file to require PHP version 8.2 instead of the previous range of 7.3 to 8.0. This change ensures that the project is using the latest stable version of PHP, improving compatibility and performance.

These updates align the project with the upgraded PHP version and Laravel framework version 11.